Move wic lib file to libs/win32
