+++ pyglet/libs/win32/wincodec.py
